randomized exponential wait in range (0, 2^x * multiplier)
Exponential backoff has to be randomized in the range of (0, 2^x * multiplier) (subject to some min and max) to properly adapt to an unknown number of contending clients attempting to access a shared resource.   The existing linear random min/max  and jitter mechanisms do not adequately address the common situation of an unknown number of contenders.   This one-line change modifies the existing exponential logic to be more correct across all common uses of exponential backoff.